moved spectrum to untils
