+++ threeML/plugins/SpectrumLike.py
@@ -10,6 +10,7 @@ from astromodels import clone_model
 from astromodels.core.parameter import Parameter
 from astromodels.functions.priors import Uniform_prior
 from astromodels.utils.valid_variable import is_valid_variable_name
+from threeML.plugins.spectrum.pha_spectrum import PHASpectrum
 
 from threeML.config.config import threeML_config
 from threeML.exceptions.custom_exceptions import custom_warnings, NegativeBackground
@@ -18,13 +19,12 @@ from threeML.io.plotting.light_curve_plots import channel_plot, disjoint_patch_p
 from threeML.io.rich_display import display
 from threeML.plugin_prototype import PluginPrototype
 from threeML.plugins.XYLike import XYLike
-from threeML.plugins.spectrum.binned_spectrum import BinnedSpectrum, ChannelSet
-from threeML.plugins.spectrum.pha_spectrum import PHASpectrum
 from threeML.utils.OGIP.likelihood_functions import half_chi2
 from threeML.utils.OGIP.likelihood_functions import poisson_log_likelihood_ideal_bkg
 from threeML.utils.OGIP.likelihood_functions import poisson_observed_gaussian_background
 from threeML.utils.OGIP.likelihood_functions import poisson_observed_poisson_background
 from threeML.utils.binner import Rebinner
+from threeML.utils.spectrum.binned_spectrum import BinnedSpectrum, ChannelSet
 from threeML.utils.stats_tools import Significance
 
 NO_REBIN = 1E-99
